harden RAW api: drop dead symbols, add RAW_kdf product, unify RAW_encode(count:), migrate size_t to Int
- delete dangling public API: RAW_strlen, RAW_memset, RAW_memcmp, RAW_malloc, deprecated let RAW_memcpy, generateRandomBytes + GenerateRandomBytesError, RAW_size_t, wordCountWithChecksum, RAW_encode_count_decl, RAW_compare(lhs_data_seeking:); rewire MemoryGuarded to CRAW and drop the RAW_mlock/munlock/free/sysconf shims
- expose RAW_kdf as a library product (was an unimportable target)
- make Mnemonic.Error public
- unify RAW_encode(count:) to assign semantics; RAW_encodables builder accumulates per element
- migrate the Swift API from size_t to Int (same type in Swift; zero behavior change)

+++ Sources/RAW/RAW_comparable.swift
@@ -7,12 +7,12 @@ import func CRAW.memcmp
 public protocol RAW_comparable {
 	/// the static comparable function for this type.
 	/// - returns: an integer value representing the comparison result. the result shall be 0 if the values are equal, else, a negative value if the left value is less than the right value, or a positive value if the left value is greater than the right value.
-	static func RAW_compare(lhs_data:UnsafeRawPointer, lhs_count:size_t, rhs_data:UnsafeRawPointer, rhs_count:size_t) -> Int32
+	static func RAW_compare(lhs_data:UnsafeRawPointer, lhs_count:Int, rhs_data:UnsafeRawPointer, rhs_count:Int) -> Int32
 }
 
 extension RAW_comparable {
 	// lexi sort is applied to the data
-	public static func RAW_compare(lhs_data:UnsafeRawPointer, lhs_count:size_t, rhs_data:UnsafeRawPointer, rhs_count:size_t) -> Int32 {
+	public static func RAW_compare(lhs_data:UnsafeRawPointer, lhs_count:Int, rhs_data:UnsafeRawPointer, rhs_count:Int) -> Int32 {
 		let result = memcmp(lhs_data, rhs_data, min(lhs_count, rhs_count))
 		if result != 0 {
 			return result
