+++ Sources/RAW/RAW_hasher.swift
@@ -6,7 +6,7 @@ import Darwin
 /// a protocol that represents a hashing algorithm.
 public protocol RAW_hasher {
 	/// the block size of the hasher.
-	static var RAW_hasher_blocksize:size_t { get }
+	static var RAW_hasher_blocksize:Int { get }
 
 	/// the output type of the hasher.
 	associatedtype RAW_hasher_outputtype:RAW_staticbuff & RAW_accessible_mutable
@@ -18,7 +18,7 @@ public protocol RAW_hasher {
 	/// update the hasher with new data from an UnsafeBufferPointer<UInt8>.
 	mutating func update(_ :UnsafeBufferPointer<UInt8>) throws
 	/// update the hasher with new data with the specified data and length arguments.
-	mutating func update(_ :UnsafeRawPointer, count:size_t) throws
+	mutating func update(_ :UnsafeRawPointer, count:Int) throws
 	/// finish a hasher by outputting to a pointer.
 	mutating func finish(into _:UnsafeMutableRawPointer) throws
 }
@@ -29,7 +29,7 @@ extension RAW_hasher {
 		try update(UnsafeRawBufferPointer(inputData))
 	}
 
-	public mutating func update(_ ptr:UnsafeRawPointer, count:size_t) throws {
+	public mutating func update(_ ptr:UnsafeRawPointer, count:Int) throws {
 		try update(UnsafeRawBufferPointer(start:ptr, count:count))
 	}
 }
