+++ Sources/RAW_sha1/SHA1.swift
@@ -7,7 +7,7 @@ public struct Hash:Sendable{}
 
 /// a SHA1 hasher.
 public struct Hasher<RAW_hasher_outputtype:RAW_staticbuff>:RAW_hasher where RAW_hasher_outputtype.RAW_fixed_type == (UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8, UInt8) {
-	public static var RAW_hasher_blocksize:size_t { size_t(__CRAWDOG_SHA1_BLOCK_SIZE) }
+	public static var RAW_hasher_blocksize:Int { Int(__CRAWDOG_SHA1_BLOCK_SIZE) }
 	
 	public typealias RAW_hasher_outputtype = Hash
 
@@ -25,7 +25,7 @@ public struct Hasher<RAW_hasher_outputtype:RAW_staticbuff>:RAW_hasher where RAW_
 		__crawdog_sha1_update(&context, buffer.baseAddress!, UInt32(buffer.count))
 	}
 		
-	public mutating func update(_ data:UnsafeRawPointer, count:size_t) {
+	public mutating func update(_ data:UnsafeRawPointer, count:Int) {
 		__crawdog_sha1_update(&context, data, UInt32(count))
 	}
